+++ rollup.config.js
@@ -21,9 +21,6 @@ preprocessConfig.markup = async function () {
   // remove whitespace
   res.code = res.code.replace(/([>}])\s+([<{])/sg, '$1$2')
 
-  // remove data-testid (only used for testing-library)
-  res.code = res.code.replace(/data-testid=".*?"/g, '')
-
   return res
 }
 
@@ -52,7 +49,11 @@ const baseConfig = {
     }),
     // make the svelte output slightly smaller
     replace({
-      'options.hydrate': 'false',
+      'options.anchor': 'undefined',
+      'options.context': 'undefined',
+      'options.customElement': 'undefined',
+      'options.hydrate': 'undefined',
+      'options.intro': 'undefined',
       delimiters: ['', ''],
       preventAssignment: true
     }),