+++ src/database/utils/extractTokens.js
@@ -1,14 +1,14 @@
 // list of emoticons that don't match a simple \W+ regex
 // extracted using:
-// require('emojibase-data/en/data.json').map(_ => _.emoticon).filter(Boolean).filter(_ => !/^\W+$/.test(_))
+// require('emoji-picker-element-data/en/emojibase/data.json').map(_ => _.emoticon).filter(Boolean).filter(_ => !/^\W+$/.test(_))
 const irregularEmoticons = new Set([
-  ':D', 'xD', ":'D", 'o:)',
-  ':x', ':p', ';p', 'xp',
-  ':l', ':z', ':j', '8D',
-  'xo', '8)', ':B', ':o',
-  ':s', ":'o", 'Dx', 'x(',
-  'D:', ':c', '>0)', ':3',
-  '</3', '<3', '\\m/', ':E',
+  ':D', 'XD', ":'D", 'O:)',
+  ':X', ':P', ';P', 'XP',
+  ':L', ':Z', ':j', '8D',
+  'XO', '8)', ':B', ':O',
+  ':S', ":'o", 'Dx', 'X(',
+  'D:', ':C', '>0)', ':3',
+  '</3', '<3', '\\M/', ':E',
   '8#'
 ])